rename to "socket" and change constructor
