+++ packages/magicbell-js/package.json
@@ -26,14 +26,14 @@
         "default": "./dist/commonjs/project-client.js"
       }
     },
-    "./realtime": {
+    "./socket": {
       "import": {
-        "types": "./dist/esm/realtime.d.ts",
-        "default": "./dist/esm/realtime.js"
+        "types": "./dist/esm/socket.d.ts",
+        "default": "./dist/esm/socket.js"
       },
       "require": {
-        "types": "./dist/commonjs/realtime.d.ts",
-        "default": "./dist/commonjs/realtime.js"
+        "types": "./dist/commonjs/socket.d.ts",
+        "default": "./dist/commonjs/socket.js"
       }
     },
     "./user-client": {
@@ -54,10 +54,11 @@
     "README.md",
     "/project-client/package.json",
     "/user-client/package.json",
-    "/realtime/package.json"
+    "/realtime/package.json",
+    "/socket/package.json"
   ],
   "scripts": {
-    "build": "run-s build:bundle build:node10 build:docs build:attw",
+    "build": "run-s build:bundle build:node10 build:attw",
     "build:attw": "attw -P --profile node16",
     "build:bundle": "tshy",
     "build:docs": "./scripts/liblab-docs.ts",
