+++ packages/magicbell-js/realtime/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "../dist/commonjs/realtime.js",
+  "types": "../dist/commonjs/realtime.d.ts"
+}
